requirements.txt
akshare
TA-Lib
mplfinance
baostock
sqlalchemy
pymysql
cryptography
backtrader
pandas~=2.0.2
starrocks
myTT
torch
scikit-learn
xgboost
transformers
lightgbm
gluonts[torch]
accelerate
evaluate
ta_cn[cn]
fire
nacos-sdk-python